touch up 64bit build cleanliness
Some of the types being used happened to work on 32bit systems because
size_t/int64 are typedefed as long long.  With a 64bit system, this is
no longer true (they just use long), so we see build errors like so:

backlight_controller.cc: In member function
  'bool power_manager::BacklightController::Init()':
backlight_controller.cc:98: error:
  no matching function for call to 'max(long long int, const long int&)'

inotify.cc: In static member function
  'static gboolean power_manager::Inotify::CallbackHandler(GIOChannel*, GIOCondition, void*)':
inotify.cc:85: error:
  cannot convert 'guint*' to 'gsize*' for argument '4' to
  'GIOError g_io_channel_read(GIOChannel*, gchar*, gsize, gsize*)'

So tweak the code to use the proper matching types rather than ones
that happen to be the same definition behind the scenes.

While we're here, also fix kInotifyBufferSize to use the right size
even though there is no explicit warning about this variable.  The
func that this gets passed to takes a "gsize", not an "int64".

BUG=chromium-os:20636
TEST=`emerge-amd64-generic power_manager` now works
TEST=`emerge-x86-generic power_manager` still works
TEST=build+booting x86-alex still works

Reviewed-on: http://gerrit.chromium.org/gerrit/8049
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>

+++ power_manager/backlight_controller.cc
@@ -95,7 +95,8 @@ bool BacklightController::Init() {
     // If there are fewer steps than the max, adjust for it.
     // TODO(sque): this is not ideal for some cases, such as max=17, where the
     // steps will have a large spread.  Something we can work on in the future.
-    num_steps_ = std::max(1LL, std::min(kMaxBrightnessSteps, max_));
+    num_steps_ = std::max(static_cast<int64>(1),
+                          std::min(kMaxBrightnessSteps, max_));
 
     // Make sure the min-max brightness range is valid.
     CHECK(max_percent_ - min_percent_ > 0);
